1.完成商家管理系统的商品分类功能（查看分类（面包屑导航），添加分类（记住父分类ID），删除分类，修改分类）

+++ pinyougou-sellergoods-service/src/main/java/com/lsf/pinyougou/sellergoods/service/impl/ItemCatServiceImpl.java
@@ -1,8 +1,11 @@
 package com.lsf.pinyougou.sellergoods.service.impl;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 
 import com.github.pagehelper.PageInfo;
+import com.lsf.pinyougou.pojogroup.TbItemCats;
 import com.lsf.pinyougou.sellergoods.service.ItemCatService;
 import org.springframework.beans.factory.annotation.Autowired;
 import com.alibaba.dubbo.config.annotation.Service;
@@ -13,7 +16,7 @@ import vo.PageResult;
 
 
 /**
- * 服务实现层
+ * 商品分类服务实现层
  *
  * @author Administrator
  */
@@ -61,7 +64,7 @@ public class ItemCatServiceImpl implements ItemCatService {
 
 
     /**
-     * 添加
+     * 添加商品分类
      */
     @Override
     public void add(TbItemCat itemCat) {
@@ -85,19 +88,41 @@ public class ItemCatServiceImpl implements ItemCatService {
      * @return
      */
     @Override
-    public TbItemCat findOne(long id) {
+    public TbItemCats findOne(long id) {
         return tbItemCatDao.queryById(id);
     }
 
 
     /**
-     * 批量删除
+     * 批量删除商品分类
      */
     @Override
-    public void batchDelete(Long[] ids) {
+    public void batchDelete(Long[] ids) throws Exception {
         for (Long id : ids) {
-            tbItemCatDao.deleteById(id);
+            // 判断分类 ID 下是否还有分类，有的话不能删除
+            TbItemCat tbItemCat = new TbItemCat();
+            tbItemCat.setParentId(id);
+            List<TbItemCat> tbItemCats = tbItemCatDao.queryAll(tbItemCat);
+            if (tbItemCats != null && tbItemCats.size() > 0) {
+                throw new Exception();
+            } else {
+                tbItemCatDao.deleteById(id);
+            }
         }
+    }
+
+
+    /**
+     * 根据上级 ID 查询商品分类
+     *
+     * @param parentId
+     * @return
+     */
+    @Override
+    public List<TbItemCat> findByParentId(Long parentId) {
+        TbItemCat tbItemCat = new TbItemCat();
+        tbItemCat.setParentId(parentId);
+        return tbItemCatDao.queryAll(tbItemCat);
     }
 
 }
